+++ src/components/error/Alert/Alert.js
@@ -1,0 +1,9 @@
+export default {
+  name: 'Alert',
+
+  props: {
+    msg: {
+      type: String,
+    },
+  },
+};
