+++ src/components/error/Alert/Alert.sass
@@ -1,0 +1,4 @@
+@import ../../../assets/css/library.sass
+
+#alert
+  background-color: $red
