+++ src/components/nav/NavBar/NavBar.sass
@@ -1,0 +1,22 @@
+@import ../../../assets/css/library.sass
+
+.hackergotchi
+  padding: .6rem
+  background-color: $blue-dark
+  border-radius: 9999px
+  color: $text-light
+  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)
+
+.iconPill
+  background-color: $white
+  padding: .6rem .3rem .7rem .8rem
+  border-top-left-radius: 9999px
+  border-bottom-left-radius: 9999px
+  box-shadow: 6px 4px 6px -1px rgba(0, 0, 0, 0.1), 6px 2px 4px -1px rgba(0, 0, 0, 0.06)
+
+.labelPill
+  background-color: $white
+  padding: .5rem .8rem .5rem .3rem
+  border-top-right-radius: 9999px
+  border-bottom-right-radius: 9999px
+  box-shadow: 4px 4px 6px -1px rgba(0, 0, 0, 0.1), 4px 2px 4px -1px rgba(0, 0, 0, 0.06)
