+++ src/components/nav/SideBar/SideBar.sass
@@ -1,0 +1,26 @@
+@import ../../../assets/css/library.sass
+
+#sidebar
+  flex-direction: column
+  background-color: $blue
+  color: $text-light
+
+.route
+  margin: 1rem
+  padding: 1rem
+  &:hover
+    box-shadow: 0 4px 6px -1px rgba(94, 117, 246, 0.3), 0 2px 4px -1px rgba(94, 117, 246, 0.08)
+    background-color: $white
+    color: $text
+
+.active
+  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)
+  background-color: $gray
+  color: $text
+
+.sysIcon
+  margin: 1rem
+  padding: 1rem
+  &:hover
+    border-radius: .5rem
+    background-color: $blue-dark
