+++ src/components/order/OrderCard/OrderCard.sass
@@ -1,0 +1,14 @@
+@import ../../../assets/css/library.sass
+
+.cardContent
+  p
+    color: $gray-dark
+  &:hover
+    cursor: pointer
+    box-shadow: 0 4px 6px -1px rgba(94, 117, 246, 0.3), 0 2px 4px -1px rgba(94, 117, 246, 0.08)
+
+.cardActive
+    border-left: 7px solid #75F65E
+
+.cardInactive
+    border-left: 7px solid #F65E75
