+++ src/components/pantry/ProductItem/ProductItem.sass
@@ -1,0 +1,2 @@
+.existence
+  margin-top: -20px
