+++ src/components/request/CreateRequestForm/CreateRequestForm.js
@@ -1,0 +1,62 @@
+import { AlertMixin } from '@/mixins/AlertMixin';
+
+export default {
+  name: 'CreateRequestForm',
+
+  mixins: [AlertMixin],
+
+  data() {
+    return {
+      description: '',
+      quantity: '',
+    };
+  },
+
+  inject: ['me'],
+
+  methods: {
+    cleanVars() {
+      this.description = '';
+      this.quantity = '';
+      this.error = null;
+    },
+
+    async createRequest() {
+      try {
+        await this.$apollo.mutate({
+          mutation: require('@/graphql/mutations/CreateRequest').default,
+          variables: {
+            input: {
+              description: this.description,
+              quantity: this.quantity,
+              order_id: this.$route.params.id,
+            },
+          },
+          update: (store, { data: { createRequest } }) => {
+            const query = {
+              query: require('@/graphql/queries/CurrentOrder').default,
+              variables: { id: this.$route.params.id },
+            };
+
+            const data = store.readQuery(query);
+            data.order.requests.push({
+              ...createRequest,
+              user: {
+                ...this.me,
+              },
+            });
+
+            store.writeQuery({
+              ...query,
+              data,
+            });
+          },
+        });
+
+        this.cleanVars();
+      } catch (e) {
+        this.error = e.message;
+      }
+    },
+  },
+};
