+++ src/components/request/RequestItem/RequestItem.sass
@@ -1,0 +1,19 @@
+@import ../../../assets/css/library.sass
+
+.bookmark
+  padding: 1.2rem 1rem 1rem 1rem
+  background-color: $gray-shadow
+  color: $text-light
+  i
+    background-color: $gray-dark
+
+.approve
+    background-color: $green-light
+    color: $text-light
+    i
+      background-color: $green
+
+.reject
+  background-color: $red-light
+  i
+    background-color: $red
