+++ src/components/user/UserCard/UserCard.sass
@@ -1,0 +1,17 @@
+@import ../../../assets/css/library.sass
+
+.cardContent
+  border: 1px solid $gray-shadow
+  p
+    color: $gray-dark
+  &:hover
+    box-shadow: 0 4px 6px -1px rgba(94, 117, 246, 0.3), 0 2px 4px -1px rgba(94, 117, 246, 0.08)
+
+.userBadge
+  background-color: $blue-dark
+
+.typePill
+  color: $text
+  background-color: $blue-light
+  p
+    color: $text
