--- src/mixins/ToggleMixin.js
@@ -1,12 +1,0 @@
-export const ToggleMixin = {
-  data() {
-    return {
-      updateForm: true,
-    };
-  },
-  methods: {
-    toggleVar(varName) {
-      this[varName] = !this[varName];
-    },
-  },
-};
