+++ src/router.js
@@ -2,13 +2,13 @@ import Vue from 'vue';
 import Router from 'vue-router';
 import { AUTH_TOKEN } from './vue-apollo';
 
-import Home from '@/views/Home';
-import Login from '@/views/Login';
-import Order from '@/views/Order';
-import Orders from '@/views/Orders';
-import Pantry from '@/views/Pantry';
-import Register from '@/views/Register';
-import Users from '@/views/Users';
+import Home from '@/views/Home/Home.vue';
+import Login from '@/views/Login/Login.vue';
+import Order from '@/views/Order/Order.vue';
+import Orders from '@/views/Orders/Orders.vue';
+import Pantry from '@/views/Pantry/Pantry.vue';
+import Register from '@/views/Register/Register.vue';
+import Users from '@/views/Users/Users.vue';
 
 Vue.use(Router);
